requirements.txt
streamlit
langchain>=0.1.0
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv

